+++ WebApplication/src/helpers/resource-helpers.js
@@ -49,6 +49,7 @@ export default {
             for (let viewer of store.state.configuration.uiOptions.ViewersOrdering) {
                 if ((["osimis-web-viewer", "stone-webviewer", "volview", "wsi"].indexOf(viewer) != -1 && viewer in store.state.configuration.installedPlugins) ||
                     (viewer.startsWith("ohif") && viewer in store.state.configuration.installedPlugins) ||
+                    (viewer.startsWith("ohif") && store.state.configuration.uiOptions.EnableOpenInOhifViewer3) ||
                     (viewer == "meddream" && store.state.configuration.uiOptions.EnableOpenInMedDreamViewer))
                 {
                     return this.getViewerUrl(level, orthancId, dicomId, viewer);